+++ extensions/ql-vscode/src/model-editor/method-modeling/method-modeling-view-provider.ts
@@ -14,10 +14,7 @@ import { assertNever } from "../../common/helpers-pure";
 import { ModelEditorViewTracker } from "../model-editor-view-tracker";
 import { ModelConfigListener } from "../../config";
 import { DatabaseItem } from "../../databases/local-databases";
-import {
-  convertFromLegacyModeledMethod,
-  convertToLegacyModeledMethod,
-} from "../shared/modeled-methods-legacy";
+import { convertFromLegacyModeledMethod } from "../shared/modeled-methods-legacy";
 
 export class MethodModelingViewProvider extends AbstractWebviewViewProvider<
   ToMethodModelingMessage,
@@ -77,9 +74,7 @@ export class MethodModelingViewProvider extends AbstractWebviewViewProvider<
         await this.postMessage({
           t: "setSelectedMethod",
           method: selectedMethod.method,
-          modeledMethod: convertToLegacyModeledMethod(
-            selectedMethod.modeledMethods,
-          ),
+          modeledMethods: selectedMethod.modeledMethods,
           isModified: selectedMethod.isModified,
         });
       }
@@ -149,29 +144,23 @@ export class MethodModelingViewProvider extends AbstractWebviewViewProvider<
       return;
     }
 
-    const views = this.editorViewTracker.getViews(
+    const view = this.editorViewTracker.getView(
       this.databaseItem.databaseUri.toString(),
     );
-    if (views.length === 0) {
-      return;
-    }
-
-    await Promise.all(views.map((view) => view.revealMethod(method)));
+    await view?.revealMethod(method);
   }
 
   private registerToModelingStoreEvents(): void {
     this.push(
       this.modelingStore.onModeledMethodsChanged(async (e) => {
-        if (this.webviewView && e.isActiveDb) {
-          const modeledMethods = e.modeledMethods[this.method?.signature ?? ""];
+        if (this.webviewView && e.isActiveDb && this.method) {
+          const modeledMethods = e.modeledMethods[this.method.signature];
           if (modeledMethods) {
-            const modeledMethod = convertToLegacyModeledMethod(modeledMethods);
-            if (modeledMethod) {
-              await this.postMessage({
-                t: "setModeledMethod",
-                method: modeledMethod,
-              });
-            }
+            await this.postMessage({
+              t: "setMultipleModeledMethods",
+              methodSignature: this.method.signature,
+              modeledMethods,
+            });
           }
         }
       }),
@@ -198,7 +187,7 @@ export class MethodModelingViewProvider extends AbstractWebviewViewProvider<
           await this.postMessage({
             t: "setSelectedMethod",
             method: e.method,
-            modeledMethod: convertToLegacyModeledMethod(e.modeledMethods),
+            modeledMethods: e.modeledMethods,
             isModified: e.isModified,
           });
         }