+++ extensions/ql-vscode/src/view/method-modeling/__tests__/MultipleModeledMethodsPanel.spec.tsx
@@ -14,7 +14,7 @@ describe(MultipleModeledMethodsPanel.name, () => {
     reactRender(<MultipleModeledMethodsPanel {...props} />);
 
   const method = createMethod();
-  const onChange = jest.fn();
+  const onChange = jest.fn<void, [ModeledMethod[]]>();
 
   describe("with no modeled methods", () => {
     const modeledMethods: ModeledMethod[] = [];
@@ -51,6 +51,23 @@ describe(MultipleModeledMethodsPanel.name, () => {
       ).toBeDisabled();
       expect(screen.queryByText("0/0")).not.toBeInTheDocument();
       expect(screen.queryByText("1/0")).not.toBeInTheDocument();
+    });
+
+    it("cannot add or delete modeling", () => {
+      render({
+        method,
+        modeledMethods,
+        onChange,
+      });
+
+      expect(
+        screen
+          .getByLabelText("Delete modeling")
+          .getElementsByTagName("input")[0],
+      ).toBeDisabled();
+      expect(
+        screen.getByLabelText("Add modeling").getElementsByTagName("input")[0],
+      ).toBeDisabled();
     });
   });
 
@@ -96,6 +113,46 @@ describe(MultipleModeledMethodsPanel.name, () => {
         screen.getByLabelText("Next modeling").getElementsByTagName("input")[0],
       ).toBeDisabled();
       expect(screen.queryByText("1/1")).not.toBeInTheDocument();
+    });
+
+    it("cannot delete modeling", () => {
+      render({
+        method,
+        modeledMethods,
+        onChange,
+      });
+
+      expect(
+        screen
+          .getByLabelText("Delete modeling")
+          .getElementsByTagName("input")[0],
+      ).toBeDisabled();
+    });
+
+    it("can add modeling", async () => {
+      render({
+        method,
+        modeledMethods,
+        onChange,
+      });
+
+      await userEvent.click(screen.getByLabelText("Add modeling"));
+
+      expect(onChange).toHaveBeenCalledWith([
+        ...modeledMethods,
+        {
+          signature: method.signature,
+          packageName: method.packageName,
+          typeName: method.typeName,
+          methodName: method.methodName,
+          methodParameters: method.methodParameters,
+          type: "none",
+          input: "",
+          output: "",
+          kind: "",
+          provenance: "manual",
+        },
+      ]);
     });
   });
 
@@ -193,6 +250,106 @@ describe(MultipleModeledMethodsPanel.name, () => {
           name: "Model type",
         }),
       ).toHaveValue("source");
+    });
+
+    it("can update the first modeling", async () => {
+      render({
+        method,
+        modeledMethods,
+        onChange,
+      });
+
+      const modelTypeDropdown = screen.getByRole("combobox", {
+        name: "Model type",
+      });
+
+      await userEvent.selectOptions(modelTypeDropdown, "source");
+
+      expect(onChange).toHaveBeenCalledWith([
+        {
+          signature: method.signature,
+          packageName: method.packageName,
+          typeName: method.typeName,
+          methodName: method.methodName,
+          methodParameters: method.methodParameters,
+          type: "source",
+          input: "Argument[this]",
+          output: "ReturnValue",
+          kind: "value",
+          provenance: "manual",
+        },
+        ...modeledMethods.slice(1),
+      ]);
+    });
+
+    it("can update the second modeling", async () => {
+      render({
+        method,
+        modeledMethods,
+        onChange,
+      });
+
+      await userEvent.click(screen.getByLabelText("Next modeling"));
+
+      const modelTypeDropdown = screen.getByRole("combobox", {
+        name: "Model type",
+      });
+
+      await userEvent.selectOptions(modelTypeDropdown, "sink");
+
+      expect(onChange).toHaveBeenCalledWith([
+        ...modeledMethods.slice(0, 1),
+        {
+          signature: method.signature,
+          packageName: method.packageName,
+          typeName: method.typeName,
+          methodName: method.methodName,
+          methodParameters: method.methodParameters,
+          type: "sink",
+          input: "Argument[this]",
+          output: "ReturnValue",
+          kind: "value",
+          provenance: "manual",
+        },
+      ]);
+    });
+
+    it("can delete modeling", async () => {
+      render({
+        method,
+        modeledMethods,
+        onChange,
+      });
+
+      await userEvent.click(screen.getByLabelText("Delete modeling"));
+
+      expect(onChange).toHaveBeenCalledWith(modeledMethods.slice(1));
+    });
+
+    it("can add modeling", async () => {
+      render({
+        method,
+        modeledMethods,
+        onChange,
+      });
+
+      await userEvent.click(screen.getByLabelText("Add modeling"));
+
+      expect(onChange).toHaveBeenCalledWith([
+        ...modeledMethods,
+        {
+          signature: method.signature,
+          packageName: method.packageName,
+          typeName: method.typeName,
+          methodName: method.methodName,
+          methodParameters: method.methodParameters,
+          type: "none",
+          input: "",
+          output: "",
+          kind: "",
+          provenance: "manual",
+        },
+      ]);
     });
   });
 
@@ -307,6 +464,102 @@ describe(MultipleModeledMethodsPanel.name, () => {
           name: "Kind",
         }),
       ).toHaveValue("remote");
+    });
+  });
+
+  describe("with 1 modeled and 1 unmodeled method", () => {
+    const modeledMethods = [
+      createModeledMethod({
+        ...method,
+        type: "sink",
+        input: "Argument[this]",
+        output: "",
+        kind: "path-injection",
+      }),
+      createModeledMethod({
+        ...method,
+        type: "none",
+        input: "",
+        output: "",
+        kind: "",
+      }),
+    ];
+
+    it("cannot add modeling", () => {
+      render({
+        method,
+        modeledMethods,
+        onChange,
+      });
+
+      expect(
+        screen.getByLabelText("Add modeling").getElementsByTagName("input")[0],
+      ).toBeDisabled();
+    });
+
+    it("can delete first modeling", async () => {
+      render({
+        method,
+        modeledMethods,
+        onChange,
+      });
+
+      await userEvent.click(screen.getByLabelText("Delete modeling"));
+
+      expect(onChange).toHaveBeenCalledWith(modeledMethods.slice(1));
+    });
+
+    it("can delete second modeling", async () => {
+      render({
+        method,
+        modeledMethods,
+        onChange,
+      });
+
+      await userEvent.click(screen.getByLabelText("Next modeling"));
+      await userEvent.click(screen.getByLabelText("Delete modeling"));
+
+      expect(onChange).toHaveBeenCalledWith(modeledMethods.slice(0, 1));
+    });
+
+    it("can add modeling after deleting second modeling", async () => {
+      const { rerender } = render({
+        method,
+        modeledMethods,
+        onChange,
+      });
+
+      await userEvent.click(screen.getByLabelText("Next modeling"));
+      await userEvent.click(screen.getByLabelText("Delete modeling"));
+
+      expect(onChange).toHaveBeenCalledWith(modeledMethods.slice(0, 1));
+
+      rerender(
+        <MultipleModeledMethodsPanel
+          method={method}
+          modeledMethods={modeledMethods.slice(0, 1)}
+          onChange={onChange}
+        />,
+      );
+
+      onChange.mockReset();
+      await userEvent.click(screen.getByLabelText("Add modeling"));
+
+      expect(onChange).toHaveBeenCalledWith([
+        ...modeledMethods.slice(0, 1),
+        {
+          signature: method.signature,
+          packageName: method.packageName,
+          typeName: method.typeName,
+          methodName: method.methodName,
+          methodParameters: method.methodParameters,
+          type: "none",
+          input: "",
+          output: "",
+          kind: "",
+          provenance: "manual",
+        },
+      ]);
     });
   });
 });