+++ extensions/ql-vscode/src/view/model-editor/__tests__/LibraryRow.spec.tsx
@@ -22,6 +22,7 @@ describe(LibraryRow.name, () => {
     showLlmButton: false,
     showMultipleModels: false,
     extensionPack: createMockExtensionPack(),
+    sourceArchiveAvailable: true,
   };
 
   const render = (props: Partial<LibraryRowProps> = {}) =>
@@ -30,15 +31,17 @@ describe(LibraryRow.name, () => {
         title="sql2o"
         libraryVersion="1.6.0"
         methods={[method]}
-        modeledMethods={{
-          [method.signature]: {
-            ...method,
-            type: "sink",
-            input: "Argument[0]",
-            output: "",
-            kind: "jndi-injection",
-            provenance: "df-generated",
-          },
+        modeledMethodsMap={{
+          [method.signature]: [
+            {
+              ...method,
+              type: "sink",
+              input: "Argument[0]",
+              output: "",
+              kind: "jndi-injection",
+              provenance: "df-generated",
+            },
+          ],
         }}
         modifiedSignatures={new Set([method.signature])}
         inProgressMethods={new InProgressMethods()}